# docs/swagger.yaml
host: 127.0.0.1:8000
info:
  contact:
    name: 黄鹏举
    url: https://github.com/huangpengju/go-gin-example
  description: |-
    实现了 Tag 的增删改查
    实现了 Article 的增删改查
    实现了 token 签发和鉴权
  termsOfService: http://swagger.io/terms/
  title: Gin 搭建 Blog API's
  version: "1.0"
paths:
  /api/v1/tags:
    get:
      parameters:
      - description: Name（标签名）
        in: query
        name: name
        type: string
      - description: State（标签状态）
        in: query
        name: state
        type: integer
      - description: Page（第几页）
        in: query
        name: page
        type: integer
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取多个标签数据
      tags:
      - Tag
  /auth:
    get:
      consumes:
      - application/json
      parameters:
      - description: 账户
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取 auth 授权（验证用户名和密码）
      tags:
      - Auth
  /tags:
    post:
      consumes:
      - application/json
      description: AddTag 接口实现了添加一个标签的功能
      parameters:
      - description: Name（标签名）
        in: query
        name: name
        required: true
        type: string
      - description: State（标签状态）
        in: query
        name: state
        type: integer
      - description: CreatedBy（创建人）
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章的标签
      tags:
      - Tag
swagger: "2.0"